Get network info from nodes in structured way
When building the cluster on ansible side, we want
to be albe to provide internal ips and hostname of
the nodes as it is on the cloud side done through
terraform.

This patch changes the output vars on terraform to provide
dicts of name:network so that we can utilize it on ansible.